vppinfra: improve bytes display in memory tracing
* Modify n_bytes/n_allocations from u32 to u64. This is
particularly helpful to track allocation beyond ~4GB

* Update the format function so that n_bytes/n_allocations
are displayed as unsigned long long integer.

* Update clib_mem_trace_sort() to avoid any underflow
during comparison

Type: improvement

+++ src/vppinfra/mem.h
@@ -225,10 +225,10 @@ typedef struct
   void *callers[12];
 
   /* Count of allocations with this traceback. */
-  u32 n_allocations;
+  u64 n_allocations;
 
   /* Count of bytes allocated with this traceback. */
-  u32 n_bytes;
+  u64 n_bytes;
 
   /* Offset of this item */
   uword offset;
